Implement preliminary campaign list with redux axios fetch
