+++ package.json
@@ -7,6 +7,7 @@
     "test": "jest"
   },
   "dependencies": {
+    "axios": "^0.17.1",
     "firebase": "^4.8.1",
     "react": "16.0.0",
     "react-native": "0.51.0",
