+++ src/components/CampaignList.js
@@ -1,21 +1,37 @@
 import React, { Component } from 'react'
-import { View, Text } from 'react-native'
+import { View, Text, FlatList } from 'react-native'
 import { connect } from 'react-redux'
 
 import { CardSection, Button } from './common'
-import { logoutUser } from '../actions'
+import CampaignItem from './CampaignItem'
+
+import { logoutUser, campaignsFetch } from '../actions'
 
 class CampaignList extends Component {
+  componentWillMount() {
+    this.props.campaignsFetch()
+  }
+
   onButtonPress() {
     this.props.logoutUser()
+  }
+
+  renderRow(campaign) {
+    return (
+      <CardSection>
+        <CampaignItem campaign={campaign.item} />
+      </CardSection>
+    )
   }
 
   render() {
     return (
       <View>
-        <Text>Campaign 1</Text>
-        <Text>Campaign 2</Text>
-        <Text>Campaign 3</Text>
+        <FlatList
+          data={this.props.campaigns}
+          renderItem={this.renderRow}
+          keyExtractor={(campaign) => campaign.id}
+        />
         <CardSection>
           <Button
             onPress={this.onButtonPress.bind(this)}
@@ -28,4 +44,10 @@ class CampaignList extends Component {
   }
 }
 
-export default connect(null, { logoutUser })(CampaignList)
+const mapStateToProps = state => {
+  return {
+    campaigns: state.campaigns
+  }
+}
+
+export default connect(mapStateToProps, { logoutUser, campaignsFetch })(CampaignList)
