+++ src/reducers/index.js
@@ -1,7 +1,9 @@
 import { combineReducers } from 'redux'
 
 import AuthReducer from './AuthReducer'
+import CampaignReducer from './CampaignReducer'
 
 export default combineReducers({
-  auth: AuthReducer
+  auth: AuthReducer,
+  campaigns: CampaignReducer
 })
